pyproject.toml
[project]
name = "edu-tools"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "pingfury", email = "pingfury@outlook.com" }
]
dependencies = [
    "streamlit>=1.41.1",
    "langchain>=0.3.14",
    "fastapi[standard]>=0.115.6",
    "langchain-google-vertexai>=2.0.9",
    "python-dotenv>=1.0.1",
    "langchain-google-genai>=2.0.7",
    "httpx>=0.27.0",
    "openai>=1.59.6",
    "langchain-community>=0.3.14",
    "google-generativeai>=0.8.3",
    "langchain-openai>=0.3.0",
    "cozepy>=0.10.3",
    "opendal>=0.45.14",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/edu_tools"]
